Combine specs for edit and update failed support specs now that editing is available
