+++ app/controllers/placements/schools/placements_controller.rb
@@ -3,6 +3,8 @@ class Placements::Schools::PlacementsController < Placements::ApplicationControl
   before_action :set_placement, only: %i[show remove destroy]
   before_action :set_decorated_placement, only: %i[show remove]
 
+  helper_method :edit_attribute_path, :add_provider_path, :add_mentor_path
+
   def index
     @pagy, placements = pagy(@school.placements.includes(:subject, :mentors).order("subjects.name"))
     @placements = placements.decorate
@@ -33,5 +35,17 @@ class Placements::Schools::PlacementsController < Placements::ApplicationControl
 
   def authorize_placement
     authorize @placement
+  end
+
+  def edit_attribute_path(attribute)
+    new_edit_placement_placements_school_placement_path(@school, @placement, step: attribute)
+  end
+
+  def add_provider_path
+    placements_school_partner_providers_path
+  end
+
+  def add_mentor_path
+    placements_school_mentors_path
   end
 end
